+++ src/Label305/DocxExtractor/Decorated/Extractors/DecoratedExtractor.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Label305\DocxExtractor\Decorated\Extractors;
+
+use DOMElement;
+
+interface DecoratedExtractor
+{
+    /**
+     * @param DOMElement $DOMElement
+     * @return mixed
+     */
+    public function extract(DOMElement $DOMElement);
+}
